node:http: move the destroy/throw teardown tests next to the after-headers ones
